fix: change the way mocks work
