+++ src/preset/setup.js
@@ -2,7 +2,4 @@
 import './configure';
 
 // Un-mock the things that we'll be mocking
-import './unmock';
-
-// Mock lean core components
-import './mock-components';
+import './mock-modules';
