--- src/preset/unmock.js
@@ -1,10 +1,0 @@
-import { getConfig } from '../lib';
-
-getConfig('coreComponents').forEach(component => {
-  try {
-    // try to un-mock the component
-    jest.unmock(component);
-  } catch (e) {
-    // do nothing if we can't
-  }
-});
